Create authorization error type
